Introduce a definition container injector for refactoring definition factories
